Rename ModelProvider file to match `amplify codegen models` output
